MOB-3392 [Quote Summary] Component UI and behavior - refactoring of the checkout component architecture - part 1
