+++ uisdk/src/main/java/com/karhoo/uisdk/screen/booking/booking/bookingcheckout/fragment/BookingCheckoutFragmentContract.kt
@@ -1,0 +1,8 @@
+package com.karhoo.uisdk.screen.booking.booking.bookingcheckout.fragment
+
+interface BookingCheckoutFragmentContract {
+    interface LoadingButtonListener {
+        fun onLoadingComplete()
+        fun showLoading()
+    }
+}
