+++ uisdk/src/main/java/com/karhoo/uisdk/screen/booking/booking/tripallocation/TripAllocationView.kt
@@ -23,7 +23,7 @@ import com.karhoo.uisdk.base.dialog.KarhooAlertDialogConfig
 import com.karhoo.uisdk.base.dialog.KarhooAlertDialogHelper
 import com.karhoo.uisdk.base.listener.SimpleAnimationListener
 import com.karhoo.uisdk.screen.booking.BookingActivity
-import com.karhoo.uisdk.screen.booking.booking.bookingrequest.BookingRequestActivity
+import com.karhoo.uisdk.screen.booking.booking.bookingcheckout.activity.BookingCheckoutActivity
 import com.karhoo.uisdk.screen.booking.domain.bookingrequest.BookingRequestStateViewModel
 import com.karhoo.uisdk.screen.booking.domain.bookingrequest.BookingRequestStatus
 import com.karhoo.uisdk.screen.rides.detail.RideDetailActivity
@@ -58,7 +58,7 @@ class TripAllocationView @JvmOverloads constructor(
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        waitForAllocation(data?.extras?.get(BookingRequestActivity.BOOKING_REQUEST_TRIP_INFO_KEY) as TripInfo)
+        waitForAllocation(data?.extras?.get(BookingCheckoutActivity.BOOKING_REQUEST_TRIP_INFO_KEY) as TripInfo)
     }
 
     private fun waitForAllocation(trip: TripInfo) {
